messages CRUD

+++ base/views.py
@@ -1,15 +1,23 @@
 from django.shortcuts import render, redirect
+from django.http import HttpResponse
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
 from django.contrib.auth.models import User
 from django.contrib.auth import authenticate, login, logout
-from .models import Room, Topic
+from django.contrib.auth.forms import UserCreationForm
+from .models import Room, Topic, Message
 from .forms import RoomForm
 
 
 def loginPage(request):
+    page = 'login'
+
+    if request.user.is_authenticated:
+        return redirect('index')
+
     if request.method == 'POST':
-        username = request.POST.get('username')
+        username = request.POST.get('username').lower()
         password = request.POST.get('password')
 
         try:
@@ -26,7 +34,7 @@ def loginPage(request):
             messages.error(request, 'Username or password does not exist')
 
     template_name = 'base/login_register.html'
-    context = {}
+    context = {'page': page}
 
     return render(request, template_name, context)
 
@@ -34,6 +42,26 @@ def loginPage(request):
 def logoutUser(request):
     logout(request)
     return redirect('index')
+
+
+def registerUser(request):
+    form = UserCreationForm()
+
+    if request.method == 'POST':
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            user = form.save(commit=False)
+            user.username = user.username.lower()
+            user.save()
+            login(request, user)
+            return redirect('index')
+        else:
+            messages.error(request, 'An error occurred during registration')
+
+    template_name = 'base/login_register.html'
+    context = {'form': form}
+
+    return render(request, template_name, context)
 
 
 def home(request):
@@ -60,12 +88,28 @@ def home(request):
 
 def room(request, pk):
     room = Room.objects.get(id=pk)
+    # get all the msgs specific to the above room
+    # the message is the Model name and _set.all() is set of msgs
+    room_messages = room.message_set.all().order_by('-created')
+    participants = room.participants.all()
+
+    if request.method == 'POST':
+        message = Message.objects.create(
+            user=request.user,
+            room=room,
+            body=request.POST.get('body')
+        )
+        room.participants.add(request.user)
+        return redirect('room', pk=room.id)
+
     template_name = 'base/room.html'
-    context = {'room': room}
+    context = {'room': room, 'room_messages': room_messages,
+               'participants': participants}
 
     return render(request, template_name, context)
 
 
+@login_required(login_url='login')
 def createRoom(request):
     form = RoomForm
     if request.method == 'POST':
@@ -80,9 +124,13 @@ def createRoom(request):
     return render(request, template_name, context)
 
 
+@login_required(login_url='login')
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        return HttpResponse('You are not allowd here!!')
 
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
@@ -95,8 +143,13 @@ def updateRoom(request, pk):
     return render(request, template_name, context)
 
 
+@login_required(login_url='login')
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse('You are not allowd here!!')
+
     if request.method == 'POST':
         room.delete()
         return redirect('index')
@@ -104,3 +157,19 @@ def deleteRoom(request, pk):
     template_name = 'base/delete.html'
 
     return render(request, template_name, {'obj': room})
+
+
+@login_required(login_url='login')
+def deleteMessage(request, pk):
+    message = Message.objects.get(id=pk)
+
+    if request.user != message.user:
+        return HttpResponse('You are not allowd here!!')
+
+    if request.method == 'POST':
+        message.delete()
+        return redirect('index')
+
+    template_name = 'base/delete.html'
+
+    return render(request, template_name, {'obj': message})
